+++ backend/dataall/modules/s3_datasets_shares/__init__.py
@@ -23,11 +23,11 @@ class S3DatasetsSharesApiModuleInterface(ModuleInterface):
     def __init__(self):
         from dataall.core.environment.services.environment_resource_manager import EnvironmentResourceManager
         from dataall.modules.s3_datasets_shares import api
-        from dataall.modules.s3_datasets_shares.services.managed_share_policy_service import SharePolicyService
+        from dataall.modules.s3_datasets_shares.services.s3_share_managed_policy_service import S3SharePolicyService
         from dataall.modules.s3_datasets.services.dataset_service import DatasetService
         from dataall.modules.datasets_base.services.dataset_list_service import DatasetListService
-        from dataall.modules.s3_datasets_shares.services.dataset_sharing_service import DatasetSharingService
-        from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareEnvironmentResource
+        from dataall.modules.s3_datasets_shares.services.s3_share_dataset_service import S3ShareDatasetService
+        from dataall.modules.s3_datasets_shares.db.s3_share_object_repositories import S3ShareEnvironmentResource
         from dataall.modules.shares_base.services.share_processor_manager import (
             ShareProcessorManager,
             ShareProcessorDefinition,
@@ -35,9 +35,9 @@ class S3DatasetsSharesApiModuleInterface(ModuleInterface):
         from dataall.modules.shares_base.services.shares_enums import ShareableType
         from dataall.modules.s3_datasets.db.dataset_models import DatasetTable, DatasetBucket, DatasetStorageLocation
 
-        EnvironmentResourceManager.register(ShareEnvironmentResource())
-        DatasetService.register(DatasetSharingService())
-        DatasetListService.register(DatasetSharingService())
+        EnvironmentResourceManager.register(S3ShareEnvironmentResource())
+        DatasetService.register(S3ShareDatasetService())
+        DatasetListService.register(S3ShareDatasetService())
 
         ShareProcessorManager.register_processor(
             ShareProcessorDefinition(ShareableType.Table, None, DatasetTable, DatasetTable.tableUri)
@@ -77,7 +77,7 @@ class S3DatasetsSharesAsyncHandlersModuleInterface(ModuleInterface):
         ]
 
     def __init__(self):
-        log.info('S3 Sharing handlers have been imported')
+        log.info('s3_datasets_shares handlers have been imported')
 
 
 class S3DatasetsSharesCdkModuleInterface(ModuleInterface):
@@ -89,9 +89,9 @@ class S3DatasetsSharesCdkModuleInterface(ModuleInterface):
 
     def __init__(self):
         import dataall.modules.s3_datasets_shares.cdk
-        from dataall.modules.s3_datasets_shares.services.managed_share_policy_service import SharePolicyService
+        from dataall.modules.s3_datasets_shares.services.s3_share_managed_policy_service import S3SharePolicyService
 
-        log.info('CDK module data_sharing has been imported')
+        log.info('CDK module s3_datasets_shares has been imported')
 
 
 class S3DatasetsSharesECSShareModuleInterface(ModuleInterface):
@@ -144,4 +144,4 @@ class S3DatasetsSharesECSShareModuleInterface(ModuleInterface):
             )
         )
 
-        log.info('ECS Share module s3_data_sharing has been imported')
+        log.info('ECS Share module s3_datasets_shares has been imported')
